updated attach_supporting_evidence_file, added file_exists_in_analysis,  unit test, and had to pay some tech debt for these changes
while previously the tests were passing, in reality the functionality of finding file by name did not work.

as for the filename to name change, in our system we use name and not filename.

the change of file_id to attachment_id is the same as above. it's what we use in our system.